+++ mlx_lm/models/bailing_moe_linear.py
@@ -424,7 +424,7 @@ class SparseMoeBlock(nn.Module):
     def __call__(self, x: mx.array) -> mx.array:
         topk_idx, topk_weight = self.gate(x)
         out = self.switch_mlp(x, topk_idx)
-        out = (out * topk_weight[..., None]).sum(axis=-2)
+        out = (out * topk_weight[..., None].astype(mx.float32)).sum(axis=-2).astype(out.dtype)
         if self.shared_experts is not None:
             out = out + self.shared_experts(x)
         return out
