+++ mlx_lm/models/minimax.py
@@ -189,7 +189,7 @@ class MiniMaxSparseMoeBlock(nn.Module):
         scores = scores.astype(x.dtype)
 
         y = self.switch_mlp(x, inds)
-        y = (y * scores[..., None]).sum(axis=-2)
+        y = (y * scores[..., None].astype(mx.float32)).sum(axis=-2).astype(y.dtype)
 
         if self.sharding_group is not None:
             y = mx.distributed.all_sum(y, group=self.sharding_group)
