+++ mlx_lm/models/phimoe.py
@@ -106,7 +106,7 @@ class PhiMoESparseMoeBlock(nn.Module):
         scores = mx.softmax(scores, axis=-1, precise=True)
 
         y = self.switch_mlp(x, inds)
-        y = (y * scores[..., None]).sum(axis=-2)
+        y = (y * scores[..., None].astype(mx.float32)).sum(axis=-2).astype(y.dtype)
 
         return y
